requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
Pillow>=11.0.0
